requirements.txt
amqp==5.2.0
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
billiard==4.2.0
cachetools==5.3.2
celery==5.3.6
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cligj==0.7.2
cron-descriptor==1.4.3
Django==4.2.10
django-celery-beat==2.5.0
django-crontab==0.7.1
django-ranged-response==0.2.0
django-simple-captcha==0.6.0
django-timezone-field==6.1.0
docker==7.0.0
fiona==1.9.5
geopandas==0.14.1
google-auth==2.27.0
greenlet==3.0.3
gunicorn==21.2.0
idna==3.6
kombu==5.3.4
kubernetes==29.0.0
networkx==3.2.1
numpy==1.26.2
oauthlib==3.2.2
osmnx==1.8.0
packaging==23.2
pandas==2.1.4
pillow==10.3.0
prompt-toolkit==3.0.43
psycopg2==2.9.9
pyasn1==0.5.1
pyasn1-modules==0.3.0
pyproj==3.6.1
python-crontab==3.0.0
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2023.3.post1
PyYAML==6.0.1
redis==5.0.1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
shapely==2.0.2
six==1.16.0
SQLAlchemy==2.0.25
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2023.3
urllib3==2.1.0
vine==5.1.0
wcwidth==0.2.13
websocket-client==1.7.0
